MVMO implemented. Not working yet.

+++ CEC15MVMO/src/Paramater.java
@@ -5,8 +5,8 @@ public class Paramater {
     public double scaling;
     public int n_par;
     public int n_tosave;
-    public int fs_factor_start;
-    public int fs_factor_end;
+    public double fs_factor_start;
+    public double fs_factor_end;
     public int n_random_last;
     public int n_random_ini;
     public double ratio_gute_max;
